updates to move ReturnURL logic to one place
